Move archiveFileName and executableFileName to class Metal.
